wip: 除批处理外测试完成
