+++ package.json
@@ -13,16 +13,12 @@
     "dependencies": {
         "axios": ">=0.18.1",
         "bootstrap": "^3.4.0",
-        "bootstrap-sass": "^3.3.7",
         "cross-env": "^3.2.3",
         "del": "^3.0.0",
         "file-system": "^2.2.2",
         "fs": "^0.0.1-security",
-        "icheck": "^1.0.2",
-        "jquery": "^3.2.1",
         "jsonfile": "^4.0.0",
         "lodash": "^4.17.5",
-        "select2": "^4.0.3",
         "vue": "^2.1.10",
         "vue-resource": "^1.3.3"
     },
@@ -33,6 +29,8 @@
         "coffeescript": ">= 2.0.0",
         "css-loader": "^3.6.0",
         "file-loader": "^6.0.0",
+        "jquery": "^3.5.1",
+        "select2": "^4.0.3",
         "select2-bootstrap-theme": "^0.1.0-beta.10",
         "style-loader": "^1.2.1",
         "tinymce": "^5.4.1",
