--- resources/assets/index.css
@@ -1,1 +1,0 @@
-@import "~bootstrap/dist/css/bootstrap.min.css";
