+++ resources/assets/ui-bootstrap-jquery/condition-group.js
