+++ resources/assets/ui-bootstrap-jquery/grid-batch.coffee
