+++ resources/views/scripts.blade.php
@@ -1,9 +1,9 @@
-@foreach(\Lego\Foundation\Facades\LegoAssets::scripts() as $script)
-    {!! \Collective\Html\HtmlFacade::script($script) !!}
-@endforeach
+{{--@foreach(\Lego\Foundation\Facades\LegoAssets::scripts() as $script)--}}
+{{--    {!! \Collective\Html\HtmlFacade::script($script) !!}--}}
+{{--@endforeach--}}
 
 @foreach(\Lego\UI\BootstrapJQueryUI::scripts() as $script)
-    <script src="/packages/wutongwan/lego/{{ $script }}"></script>
+    <script src="{{ $script }}"></script>
 @endforeach
 
 <script>
